refactor pm quality summary detail router
