+++ src/app/router_registry.py
@@ -125,6 +125,9 @@ from app.routers.dpm_command_center_pm_quality_review_actions import (
 from app.routers.dpm_command_center_pm_quality_score_run_lookup import (
     router as dpm_command_center_pm_quality_score_run_lookup_router,
 )
+from app.routers.dpm_command_center_pm_quality_summary_detail import (
+    router as dpm_command_center_pm_quality_summary_detail_router,
+)
 from app.routers.dpm_command_center_pm_quality_summary_invocations import (
     router as dpm_command_center_pm_quality_summary_invocations_router,
 )
@@ -396,6 +399,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(dpm_command_center_pm_quality_review_action_lookup_router)
     app.include_router(dpm_command_center_pm_quality_score_run_lookup_router)
     app.include_router(dpm_command_center_pm_quality_summary_invocations_router)
+    app.include_router(dpm_command_center_pm_quality_summary_detail_router)
     app.include_router(dpm_command_center_pm_quality_summary_lookup_router)
     app.include_router(dpm_command_center_portfolio_memory_router)
     app.include_router(dpm_command_center_portfolio_memory_search_router)
